nomeando os arquivos para a arquitetura desejada #8
