+++ accounts/router.py
@@ -1,7 +1,9 @@
 from django.urls import path
 from rest_framework_simplejwt.views import TokenRefreshView
 
-from .views import LoginView, MeView, RegisterView
+from .controller import LoginView, MeView, RegisterView
+
+app_name = "accounts"
 
 urlpatterns = [
     path("login/", LoginView.as_view(), name="auth-login"),
