+++ catalogo_depara/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import CatalogoDePara
+from .entity import CatalogoDePara
 
 
 @admin.register(CatalogoDePara)
