--- catalogo_depara/urls.py
@@ -1,12 +1,0 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-from .views import CatalogoDeparaViewSet
-
-app_name = 'catalogo_depara'
-
-router = DefaultRouter()
-router.register(r'', CatalogoDeparaViewSet, basename='catalogo_depara')
-
-urlpatterns = [
-    path('', include(router.urls)),
-]
